TRUST-407 adds duplicates for couchdb keeping original files

+++ services/DBService.js
@@ -1,8 +1,10 @@
-const {normalizeNumber, validateSort, buildSelector, DBKeys, pruneOpenDSUFields, remapObject} = require("../utils");
+const {DBKeys} = require("../utils");
 const logger = $$.getLogger("DBService");
 const nano = require("nano");
-const {OpenDSUKeys} = require("../utils/constants");
-const {processInChunks} = require("../utils/chunk");
+const {ensureAuth} = require("./utils");
+const {DatabaseClient} = require("./DBClient");
+
+let dbService;
 
 /**
  * DBService for database operations.
@@ -14,23 +16,22 @@ class DBService {
      * @param {{uri: string, username?: string, secret?: string}} config - Configuration object containing database connection details.
      */
     constructor(config) {
+        if (dbService)
+            return dbService;
+
+        dbService = this;
         this.config = config;
-        this.dbConnection = this.__createConnection(config);
+        this.client = this.__createConnection(config);
         [
             this.dbExists,
             this.createDatabase,
             this.openDatabase,
             this.deleteDatabase,
             this.listDatabases,
-            this.countDocs,
             this.addIndex,
-            this.insertDocument,
-            this.readDocument,
-            this.updateDocument,
-            this.deleteDocument,
-            this.listDocuments,
-            this.filter,
-        ].forEach(this.__ensureAuth.bind(this));
+        ].forEach((m) => ensureAuth(this, logger, m));
+
+        this.databases = {}
     }
 
     /**
@@ -39,8 +40,8 @@ class DBService {
      * @returns {nano.ServerScope} - A database client instance.
      */
     __createConnection(config) {
-        if (this.dbConnection)
-            return this.dbConnection;
+        if (this.client)
+            return this.client;
 
         const url = new URL(config.uri);
         if (url.username && url.password)
@@ -49,40 +50,14 @@ class DBService {
         const username = config.username || url.username || "";
         const password = config.secret || url.password || "";
 
-        this.dbConnection = nano({
+        this.client = nano({
             url: config.uri,
             requestDefaults: {auth: {username, password}}
         });
 
-        return this.dbConnection;
+        return this.client;
     }
 
-    /**
-     *
-     * @param {Function} method
-     * @private
-     */
-    __ensureAuth(method){
-        const name = method.name
-        const original = this[name];
-        this[name] = async function(...args){
-            try {
-                return await original.apply(this, args);
-            } catch (e){
-                if (e.statusCode === 401 || e['status-code'] === 401) {
-                    try {
-                        console.debug(`Cookie expired - Re-authenticating with CouchDB server`);
-                        await this.dbConnection.auth(this.config.username, this.config.secret);
-                    } catch (err){
-                        throw new Error(`Failed to authenticate with CouchDB server to redo the ${name} operation. Error: ${err.message || err}. Original Error: ${e.message || e}`);
-                    }
-                    return await original.apply(this, args);
-                }
-                this._testErrorForShutdown(e)
-                throw e;
-            }
-        }.bind(this)
-    }
 
     /**
      * Checks if DB Name is valid for couch db.
@@ -119,7 +94,7 @@ class DBService {
     async dbExists(dbName) {
         try {
             dbName = this.changeDBNameToLowerCaseAndValidate(dbName);
-            const dbList = await this.dbConnection.db.list();
+            const dbList = await this.client.db.list();
             return dbList.includes(dbName);
         } catch (error) {
             throw new Error(`Failed to check if database "${dbName}" exists: ${error.message || error}`);
@@ -142,7 +117,7 @@ class DBService {
                 return true;
             }
 
-            await this.dbConnection.db.create(dbName);
+            await this.client.db.create(dbName);
             logger.info(`Database "${dbName}" created successfully.`);
 
             const indexSet = new Set(Array.isArray(indexes) && indexes.length ? indexes : []);
@@ -160,19 +135,29 @@ class DBService {
      * Retrieves an existing database, or creates it if it does not exist.
      *
      * @param {string} dbName - Database name to retrieve or create.
-     * @returns {Promise<nano.DocumentScope>} A Promise that resolves to database instance.
+     * @returns {Promise<DatabaseClient>} A Promise that resolves to database instance.
      * @throws {Error} Throws an error if the database retrieval or creation process fails.
      */
     async openDatabase(dbName) {
+        const self = this;
+
         try {
             dbName = this.changeDBNameToLowerCaseAndValidate(dbName);
+
+            function openAndCache(){
+                if (!(dbName in self.databases)){
+                    self.databases[dbName] = new DatabaseClient(self.config, dbName);
+                }
+                return self.databases[dbName];
+            }
+
             if (await this.dbExists(dbName))
-                return this.dbConnection.use(dbName);
+                return openAndCache()
 
             logger.info(`Database does not exist. Creating new database "${dbName}".`);
             await this.createDatabase(dbName);
             // TODO - Remove, return DBService instance
-            return this.dbConnection.use(dbName);
+            return openAndCache()
         } catch (error) {
             throw new Error(`Error in openDatabase: ${error.message || error}`);
         }
@@ -186,7 +171,8 @@ class DBService {
     async deleteDatabase(dbName) {
         try {
             dbName = this.changeDBNameToLowerCaseAndValidate(dbName);
-            await this.dbConnection.db.destroy(dbName);
+            await this.client.db.destroy(dbName);
+            delete(this.databases[dbName]);
             return true;
         } catch (error) {
             if (error.status === 404) {
@@ -207,13 +193,13 @@ class DBService {
     async listDatabases(verbose = false) {
         const self = this;
         try {
-            const list = await this.dbConnection.db.list();
+            const list = await this.client.db.list();
             if (!verbose)
                 return list;
 
             const databaseInfoList = [];
             for (const dbName of list) {
-                const metadata = await self.dbConnection.use(dbName).info(); // Get metadata of the database
+                const metadata = await self.client.use(dbName).info(); // Get metadata of the database
                 databaseInfoList.push({
                     name: dbName,
                     type: "collection",
@@ -227,58 +213,38 @@ class DBService {
     }
 
     /**
-     * Test is the error is worth shutting down the system for
-     * @param {Error} error
-     * @returns {void}
-     */
-    _testErrorForShutdown(error){
-        if (error.message.includes("ECONNREFUSED")){
-            logger.error("Failed to connect to couchdb instance. Shutting down the system...");
-            process.exit(1);
-        }
-    }
-
-    /**
      * Retrieves the document count for a specific table
      *
-     * @param {string} tableName - The table name to get the document count
+     * @param {string} database - The table name to get the document count
      * @returns {Promise<number>} - A promise that resolves to the document count of the specified table.
      * @throws {Error} - Throws an error if retrieving the document count fails.
      */
-    async countDocs(tableName) {
-        try {
-            const info = await this.dbConnection.db.use(tableName).info();
-            return info.doc_count || 0;
-        } catch (error) {
-            if (error.statusCode === 404) {
-                logger.warn(`Table "${tableName}" does not exist. Unable to count documents.`);
-                return 0;
-            }
-            throw new Error(`Failed to retrieve document count for table ${tableName}: ${error}`);
-        }
+    async countDocs(database) {
+        const db = await this.openDatabase(database);
+        return await db.countDocs()
     }
 
     /**
      * Adds indexes to a specific table.
      *
-     * @param {string} tableName - The name of the table to add the index to.
+     * @param {string} database - The name of the table to add the index to.
      * @param {string | Array<string>} properties - The property or property list to be indexed.
      * @returns {Promise<boolean>} - Resolves to `true` if the index was successfully created.
      * @throws {Error} - Throws an error if the table doesn't exist or if adding the index fails.
      */
-    async addIndex(tableName, properties) {
+    async addIndex(database, properties) {
         if (!properties || (Array.isArray(properties) && properties.length === 0)) {
-            logger.info(`No indexes provided for table: ${tableName}. Skipping index creation.`);
+            logger.info(`No indexes provided for table: ${database}. Skipping index creation.`);
             return false;
         }
 
-        if (!await this.dbExists(tableName))
-            throw new Error(`Table "${tableName}" does not exist.`);
+        if (!await this.dbExists(database))
+            throw new Error(`Table "${database}" does not exist.`);
 
         properties = Array.isArray(properties) ? properties : [properties];
         const index = `${properties.join("_")}_index`;
         try {
-            await this.dbConnection.use(tableName).createIndex({
+            await this.client.use(database).createIndex({
                 name: index,
                 index: {
                     fields: properties
@@ -286,159 +252,84 @@ class DBService {
                 type: "json" // default
             });
 
-            logger.info(`Added index ${index} for table "${tableName}".`);
+            logger.info(`Added index ${index} for table "${database}".`);
             return true;
         } catch (err) {
-            throw new Error(`Could not add index ${index} on ${tableName}: ${err.message}`);
+            throw new Error(`Could not add index ${index} on ${database}: ${err.message}`);
         }
     }
 
     /**
      * Inserts a document into a specified table.
-     * @param {string} tableName
+     * @param {string} database
      * @param {string} _id - The primary key for the record.
      * @param {Object} document - The document to insert.
      * @returns {Promise<{ [key: string]: any }>} - The inserted document.
      * @throws {Error} - Throws an error if any operation fails, including checking if the record exists or inserting the record.
      */
-    async insertDocument(tableName, _id, document) {
+    async insertDocument(database, _id, document) {
         // TODO - Empty objects {} are not being validated.
-        try {
-            await this.openDatabase(tableName);
-            const record = await this.readDocument(tableName, _id).catch(() => undefined);
-            if (record)
-                throw new Error(`A record with PK "${_id}" already exists in ${tableName}`);
-
-            const insert = {
-                ...pruneOpenDSUFields(document),
-                [DBKeys.PK]: _id,
-                [DBKeys.TIMESTAMP]: Date.now()
-            };
-
-            const {id} = await this.dbConnection.use(tableName).insert(insert);
-            return await this.readDocument(tableName, id);
-        } catch (err) {
-            throw err;
-        }
+        const db = await this.openDatabase(database);
+        return await db.insertDocument(_id, document);
     }
 
     /**
      * Retrieves a document by its ID from the specified table.
-     * @param {string} tableName
+     * @param {string} database
      * @param {string} _id - The ID of the document to retrieve.
      * @returns {Promise<{ pk: string, [key: string]: any }>} - The retrieved document.
      */
-    async readDocument(tableName, _id) {
-        try {
-            await this.openDatabase(tableName);
-            const document = await this.dbConnection.use(tableName).get(_id);
-            return remapObject(document);
-        } catch (error) {
-            if (error.statusCode !== 404)
-                logger.error(`Failed to retrieve document ${_id} from table ${tableName}:`, error);
-            throw error;
-        }
+    async readDocument(database, _id) {
+        const db = await this.openDatabase(database);
+        return await db.readDocument(_id);
     }
 
     /**
      * Updates a record in the specified table.
      * If the record does not exist and the `fallbackInsert` flag is set to true, it will insert the record instead.
      *
-     * @param {string} tableName
+     * @param {string} database
      * @param {string} _id - The ID of the document to update.
      * @param {Object} document - The record data to update.
      * @returns {Promise<{ [key: string]: any }>} - The updated or inserted record.
      * @throws {Error} - If the operation fails.
      */
-    async updateDocument(tableName, _id, document) {
-        try {
-            const dbClient = this.dbConnection.use(tableName);
-            const dbRecord = await dbClient.get(_id);
-            const _rev = dbRecord[DBKeys.REV];
-            for (let prop in document) {
-                dbRecord[prop] = document[prop];
-            }
-
-            const update = {
-                ...pruneOpenDSUFields(dbRecord),
-                [DBKeys.PK]: _id,
-                [DBKeys.REV]: _rev,
-                [DBKeys.TIMESTAMP]: Date.now()
-            };
-
-            const response = await dbClient.insert(update);
-            return await this.readDocument(tableName, response.id);
-        } catch (error) {
-            if (error.statusCode === 404) {
-                if (document[OpenDSUKeys.FALLBACK_INSERT]) {
-                    delete document[OpenDSUKeys.FALLBACK_INSERT];
-                    return this.insertDocument(tableName, _id, document);
-                }
-                throw new Error(`Failed to update document "${_id}" from "${tableName}": Not found.`);
-            }
-
-            throw new Error(`Failed to update document "${_id}" from "${tableName}": ${error}`);
-        }
+    async updateDocument(database, _id, document) {
+        const db = await this.openDatabase(database);
+        return await db.updateDocument(_id,  document);
     }
 
     /**
      * Deletes a record from the specified table.
-     * @param {string} tableName
+     * @param {string} database
      * @param {string} _id - The primary key (ID) of the record to delete.
      * @returns {Promise<{ pk: string }>} - The deleted record.
      * @throws {Error} - If the operation fails.
      */
-    async deleteDocument(tableName, _id) {
-        try {
-            const dbClient = this.dbConnection.use(tableName);
-            const document = await dbClient.get(_id);
-            await dbClient.destroy(_id, document[DBKeys.REV]);
-            return {[OpenDSUKeys.PK]: _id};
-        } catch (error) {
-            if (error.statusCode === 404)
-                return {[OpenDSUKeys.PK]: _id};
-
-            throw new Error(`Error deleting document ${_id} from table ${tableName}: ${error}`);
-        }
+    async deleteDocument(database, _id) {
+        const db = await this.openDatabase(database);
+        return await db.deleteDocument(_id)
     }
 
     /**
      * Lists documents from a specified table.
      *
-     * @param {string} tableName - The name of the table to fetch documents from.
+     * @param {string} database - The name of the table to fetch documents from.
      * @param {Object} [options={}] - Optional configuration object for listing documents.
      * @param {number} [options.limit] - The maximum number of documents to fetch (optional).
      * @returns {Promise<Array<{ [key: string]: any }>>} - A promise that resolves to an array of document objects.
      * @throws {Error} - Throws an error if the query fails.
      */
-    async listDocuments(tableName, options = {}) {
-        const {limit} = options;
-        tableName = this.changeDBNameToLowerCaseAndValidate(tableName);
-
-        try {
-            await this.openDatabase(tableName);
-            const queryOptions = {
-                include_docs: true,
-                startkey: '',
-                endkey: '_design/',
-                inclusive_end: false // Exclude design docs
-            };
-
-            if (limit && Number.isInteger(limit) && limit > 0)
-                queryOptions.limit = limit;
-
-            const response = await this.dbConnection.use(tableName).list(queryOptions);
-            return processInChunks(response.rows, 2, (row) => remapObject(row.doc));
-        } catch (error) {
-            throw new Error(`Error listing documents from table ${tableName}: ${error}`);
-        }
+    async listDocuments(database, options = {}) {
+        const db = await this.openDatabase(database);
+        return await db.listDocuments(options);
     }
 
     /**
      * Filters documents from specified table.
      *
      * @async
-     * @param {string} tableName - The name of the table to query.
+     * @param {string} database - The name of the table to query.
      * @param {Array<string>} query - The query object to filter documents.
      * @param {Array<Object>} [sort=[]] - Sorting criteria for the results.
      * @param {number} [limit=undefined] - Maximum number of documents to return.
@@ -446,27 +337,9 @@ class DBService {
      * @returns {Promise<Array<Object>>}g.
      * @throws {Error} If there is an issue querying the database.
      */
-    async filter(tableName, query, sort = [], limit = undefined, skip = 0) {
-        tableName = this.changeDBNameToLowerCaseAndValidate(tableName);
-        limit = normalizeNumber(limit, 1, undefined);
-        skip = normalizeNumber(skip, 0, 0);
-        sort = validateSort(sort);
-
-        const selector = buildSelector(query);
-        const mangoQuery = {
-            selector,
-            // fields: [],
-            sort,
-            skip,
-            ...(limit ? {limit} : {})
-        };
-
-        try {
-            const result = await this.dbConnection.use(tableName).find(mangoQuery);
-            return processInChunks(result.docs, 2, (doc) => remapObject(doc));
-        } catch (error) {
-            throw new Error(`Error filtering documents from table ${tableName}: ${error}`);
-        }
+    async filter(database, query, sort = [], limit = undefined, skip = 0) {
+        const db = await this.openDatabase(database);
+        return await db.filter(query, sort, limit, skip);
     }
 
 }
